Remove PF2eVisionerRegionBehavior and replace with VisibilityRegionBehavior in region registration
- Deleted the PF2eVisionerRegionBehavior class from pf2e-visioner-region-behavior.js, which managed visibility states for tokens in regions.
- Updated register.js to import and register the new VisibilityRegionBehavior instead of the removed PF2eVisionerRegionBehavior.
